Finished
I checked as udacity rubric rules
