+++ projects/weather-journal-app/website/app.js
@@ -1,83 +1,68 @@
-// const { header } = require("express/lib/request");
-
-/* Global Variables */
-const url = 'https://openweathermap.org/';
-// const key = '&appid=9f15e45060...';
-const key = 'AIzaSyCnCWYDaa9mKdg6qrzKF7E9HNK_9BHyO7E';
-
+// Api url for Open weather map
+const url = 'https://api.openweathermap.org/data/2.5/weather?id=';
 
 // Personal API Key for OpenWeatherMap API
-const retrieveData = async(url, data = {}) => {
-    console.log(data);
-    const response = await fetch(url, {
-        method: 'POST',
-        credentials: 'same-origin',
-        headers: {
-            'Content-type': 'application.json',
-        },
-        body: JSON.stringify(data),
-    })
-    try {
-        const newData = await response.json();
-        console.log(newData);
-        return newData
-    } catch (error) {
-        console.log("error", error);
-        // appropriately handle the error
-    }
-}
+const apiKey = '&appid=cb2eeafd3fb2f4ce0b2db2232ed72f9b';
 
 // Event listener to add function to existing HTML DOM element
-document.querySelector('#generate').addEventListener('click', callBack)
-    /* Function called by event listener */
-function callBack(e) {
-    const zip = document.querySelector('#zip').value;
-    const feelings = document.querySelector('#feelings').value
-    getZip(url, zip, key)
-        // New Syntax!
+document.getElementById('generate').addEventListener('click', performAction);
+
+/* Function called by event listener */
+function performAction(e) {
+    const zip = document.getElementById('zip').value;
+    const response = document.getElementById('feelings').value;
+    getZip(url, zip, apiKey)
         .then(function(data) {
             // Add data
             console.log(data);
-            postData('/add', { temperature: data.temperature, date: data.date, userResponse: data.userResponse });
+            postData('/addInfo', { temperature: data.main.temp, date: data.dt, feel: response });
         })
-        .then(
-            updateUI()
-        )
-}
+        .then(updateUI())
+};
 
-const updateUI = async() => {
-    const request = await fetch('/all');
-    try {
-        const allData = await request.json();
-        document.getElementById('date').innerHTML = allData[0].date;
-        document.getElementById('temp').innerHTML = allData[0].temperature;
-        document.getElementById('content').innerHTML = allData[0].userResponse;
-
-    } catch (error) {
-        console.log("error", error);
-    }
-}
+/* Function to GET Web API Data*/
 const getZip = async(url, zip, key) => {
         const res = await fetch(url + zip + key);
 
         try {
             const newData = await res.json();
-            console.log(newData);
             return newData
         } catch (error) {
             console.log('error', error);
         }
     }
-    /* Function to GET Web API Data*/
-const getData = async(url = '', data = {}) => {
-        console.log(data);
-        const response = await fetch(url, {})
-    }
     /* Function to POST data */
+const postData = async(url = '', data = {}) => {
+    console.log(data);
+    const response = await fetch(url, {
+        method: 'POST',
+        credentials: 'same-origin',
+        headers: {
+            'Content-Type': 'application/json',
+        },
+        // Body data type must match "Content-Type" header        
+        body: JSON.stringify(data),
+    });
 
+    try {
+        const newData = await response.json();
+        console.log(newData);
+        return newData;
+    } catch (error) {
+        console.log("error", error);
+    }
+}
 
-/* Function to GET Project Data */
+/* Function to Update UI */
+const updateUI = async() => {
+    const request = await fetch('/all');
+    try {
+        const allData = await request.json();
+        document.getElementById('temp').innerHTML = Math.round(allData[0].temperature) + 'degrees';
+        document.getElementById('date').innerHTML = allData[0].date;
+        document.getElementById('content').innerHTML = allData[0].res
 
-// Create a new date instance dynamically with JS
-// let d = new Date();
-// let newDate = d.getMonth() + '.' + d.getDate() + '.' + d.getFullYear();
+    } catch (error) {
+        console.log("error", error);
+    }
+}
