--- projects/weather-journal-app/website/commentsOnlyJS/app.js
@@ -1,33 +1,0 @@
-// Personal API Key for OpenWeatherMap API
-
-// Event listener to add function to existing HTML DOM element
-
-/* Function called by event listener */
-
-/* Function to GET Web API Data*/
-
-/* Function to POST data */
-
-
-/* Function to GET Project Data */
-
-// Testing
-// const http = require("http");
-
-// const hostname = "127.0.0.1";
-// const port = 8000;
-
-// // Create HTTP server
-// const server = http.createServer(function(req, res) {
-
-//    // Set the response HTTP header with HTTP status and Content type
-//    res.writeHead(200, {'Content-Type': 'text/plain'});
-
-//    // Send the response body "Hello World"
-//    res.end('Hello World\n');
-// });
-
-// // Prints a log once the server starts listening
-// server.listen(port, hostname, function() {
-//    console.log(`Server running at http://${hostname}:${port}/`);
-// })
