+++ projects/weather-journal-app/website/index.html
@@ -14,8 +14,9 @@
             Weather Journal App
         </div>
         <div class="holder zip">
-            <label for="zip">Enter Zipcode here</label>
-            <input type="text" id="zip" placeholder="enter zip code here">
+            <label for="zip">Enter city id here</label>
+            <!-- test id = 2172797 -->
+            <input type="text" id="zip" placeholder="enter city id">
         </div>
         <div class="holder feel">
             <label for="feelings">How are you feeling today?</label>
@@ -31,7 +32,7 @@
             </div>
         </div>
     </div>
-    <script src="appTest.js" type="text/javascript"></script>
+    <script src="app.js" type="text/javascript"></script>
 
 </body>
